Avoid repeating the creation of EigenData instances (#35)
* Move creation of eigen data dict to eigenclass.py

This heretofore was being created in the NEI class, which slowed down
execution.  By moving it to eigenclass.py, it should only run once on
import.

Ultimately, this should be relegated to an HDF5 file so that the
calculations need not be repeated.  It could also be helpful to store
this information in the new data classes available from Python 3.7 on.

* Reformat file