#41 plus! (#42)
* Copy over testing workflow from PlasmaPy

* Remove tentatively added conda python package thing

* Remove Python 3.6 tests from GitHub Actions file

* Copy over tox.ini from PlasmaPy

* Remove Python 3.6 stuff from tox.ini

* Update copyright years

* Replace old pytest config with current one

The previous one referred to functionality of pytest-doctestplus,
astropy's testing plugin, which we've gotten rid of in plasmapy proper
a while ago. There was an addopts clause that added the unrecognized
flag.

* Remove obsolete azure-pipelines.yml

Fully replaced by github actions now.

* Ensure we're testing the, uh, right package.

* Add a step to pull files from Git LFS

as per https://github.com/actions/checkout/issues/165#issuecomment-657673315

* Typo fix

* Try to fix broken yml file

* Typo fix

* Quickly ensure docs tox env has access to sphinx

this is done through selecting all the extras, one of which has sphinx
in it. Hackish.

* Remove circleci

Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>